- TASK 1296: Adding Facebook messenger to the website

Related Work Items: #1296
- BUG 1287: The payout is too generous for challenge of 6 participants

Related Work Items: #1287
- BUG 1281: Transaction history display is not in order

Related Work Items: #1281
- BUG 1272 - BUG 1273

Related Work Items: #1272, #1273
- TASK 1317: Add duration property to challenge dto

Related Work Items: #1317
- Cashier challenge refactored
- Challenge aggregator refactored
- US 1098: As a user, I want to be notified by email whenever a challenge has started

Related Work Items: #1098
- TASK 1282: Add a text boxes to describe the basic rules for all challenges

Related Work Items: #1282
- Challenge frontend design
- TASK 1278 and 1280

Related Work Items: #1278, #1280
- Transaction order bug fixed
- TASK 1276

Related Work Items: #1276
- BUG 1305

Related Work Items: #1305
- BUG 1255

Related Work Items: #1255
- TASK 1279

Related Work Items: #1279
- TASK 1262

Related Work Items: #1262
- Cashier challenge participant payout refactored
- Cashier API refactoring
- BUG 1197

Related Work Items: #1197
- BUG 1209

Related Work Items: #1209
- Frontend bug fixed
- BUG 1257

Related Work Items: #1257

Related work items: #1098, #1197, #1209, #1255, #1257, #1262, #1272, #1273, #1276, #1278, #1279, #1280, #1281, #1282, #1287, #1296, #1305, #1317

+++ src/eDoxa.Games.Api/Application/Services/GameCredentialService.cs
@@ -4,6 +4,7 @@
 // ================================================
 // Copyright © 2019, eDoxa. All rights reserved.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -61,15 +62,25 @@ namespace eDoxa.Games.Api.Application.Services
 
         public async Task<IDomainValidationResult> UnlinkCredentialAsync(Credential credential)
         {
-            credential.Delete();
+            var result = new DomainValidationResult();
 
-            await _gameCredentialRepository.UnitOfWork.CommitAsync();
+            if (credential.Timestamp > DateTime.UtcNow.AddMonths(-1))
+            {
+                result.AddFailedPreconditionError($"You will have the right to unlink your {credential.Game.DisplayName} credentials in {(credential.Timestamp - DateTime.UtcNow.AddMonths(-1)).Days} days.");
+            }
 
-            _gameCredentialRepository.DeleteCredential(credential);
+            if (result.IsValid)
+            {
+                credential.Delete();
 
-            await _gameCredentialRepository.UnitOfWork.CommitAsync();
+                await _gameCredentialRepository.UnitOfWork.CommitAsync();
 
-            return new DomainValidationResult();
+                _gameCredentialRepository.DeleteCredential(credential);
+
+                await _gameCredentialRepository.UnitOfWork.CommitAsync();
+            }
+
+            return result;
         }
 
         public async Task<IReadOnlyCollection<Credential>> FetchCredentialsAsync(UserId userId)
